pyproject.toml
[tool.poetry]
name = "oncoai-prototype"
version = "1.0.0"
description = "Python application for OncoAI Patient Outcome Navigator."
authors = ["Sangeeth George sangeethgeorge1@gmail.com"]
readme = "README.md"
packages = [
    # oncoai_prototype is in the 'src' directory, so its 'from' should be "src"
    { include = "oncoai_prototype", from = "src" },
    # streamlit_app is directly in the project root, so its 'from' should be "."
    { include = "streamlit_app", from = "." }
]

[tool.poetry.dependencies]
python = "^3.11.8"
pandas = "^2.3.0"
scikit-learn = "^1.7.0"
duckdb = "^1.3.1"
matplotlib = "^3.10.3"
seaborn = "^0.13.2"
xgboost = "^3.0.2"
numpy = "2.2.6"
shap = "^0.48.0"
jupyterlab = "^4.4.4"
ipykernel = "^6.29.5"
jupyterlab-quarto = "^0.3.5"
jupyterlab-git = "^0.51.2"
ipywidgets = "^8.1.7"
jupyter-resource-usage = "^1.1.1"
streamlit = "^1.46.1"
statsmodels = "^0.14.4"
lightgbm = "^4.6.0"
pytest = "^8.4.1"
joblib = "^1.5.1"
python-dotenv = "^1.1.1"
mlflow = "^3.1.1"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.4.4"
ipykernel = "^6.29.4"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "--import-mode=importlib"
